channelTreeの仮書き換え
