+++ src/_store/domain/index.ts
@@ -1,6 +1,5 @@
 import { defineModule } from 'direct-vuex'
 import { me } from './me'
-import { channelTree } from './channelTree'
 import { messagesView } from './messagesView'
 import { stampCategory } from './stampCategory'
 
@@ -8,7 +7,6 @@ export const domain = defineModule({
   namespaced: true,
   modules: {
     me,
-    channelTree,
     messagesView,
     stampCategory
   }
