+++ src/store/domain/channelTree/state.ts
@@ -1,0 +1,15 @@
+import { ChannelTree } from '@/lib/channelTree'
+
+export interface S {
+  channelTree: Readonly<ChannelTree>
+  homeChannelTree: Readonly<ChannelTree>
+}
+
+export const state: S = {
+  channelTree: {
+    children: []
+  },
+  homeChannelTree: {
+    children: []
+  }
+}
